+++ pancake2viper/src/ir/mod.rs
@@ -1,6 +1,9 @@
 mod expression;
+pub mod mangle;
 mod statement;
+pub mod to_shape;
 mod toplevel;
+pub mod types;
 pub mod utils;
 
 pub use expression::*;
