--- pancake2viper/src/ir_to_viper/utils/mod.rs
@@ -1,20 +1,0 @@
-mod auto;
-mod to_shape;
-
-use crate::utils::{ToViper, ToViperError, TryToViper, ViperEncodeCtx};
-
-impl<'a, T: TryToViper<'a>> TryToViper<'a> for Vec<T> {
-    type Output = Vec<T::Output>;
-    fn to_viper(self, ctx: &mut ViperEncodeCtx<'a>) -> Result<Self::Output, ToViperError> {
-        self.into_iter()
-            .map(|a| a.to_viper(ctx))
-            .collect::<Result<Vec<_>, _>>()
-    }
-}
-
-impl<'a, T: ToViper<'a>> ToViper<'a> for Vec<T> {
-    type Output = Vec<T::Output>;
-    fn to_viper(self, ctx: &mut ViperEncodeCtx<'a>) -> Self::Output {
-        self.into_iter().map(|a| a.to_viper(ctx)).collect()
-    }
-}
