+++ pancake2viper/src/lib.rs
@@ -2,7 +2,6 @@ pub mod annotation;
 pub mod cli;
 pub mod ir;
 pub mod ir_to_viper;
-pub mod mangle_ir;
 pub mod pancake;
 pub mod pancake_to_ir;
 pub mod utils;
